[core] Prevent folder removal during version switching

+++ core/magisk_module/uninstall.sh
@@ -17,5 +17,14 @@
 # Copyright (C) 2021 LSPosed Contributors
 #
 
-rm -rf /data/adb/lspd
+MODDIR=${0%/*}
+MODSDIR=$(dirname "$MODDIR")
+if [ -d "$MODSDIR/riru_lsposed" ] && [ -d "$MODSDIR/zygisk_lsposed" ]; then
+  if [ -f "$MODSDIR/riru_lsposed/remove" ] && [ -f "$MODSDIR/zygisk_lsposed/remove" ]; then
+    rm -rf /data/adb/lspd
+  fi
+else
+  rm -rf /data/adb/lspd
+fi
+
 rm -rf /data/adb/riru/modules/lspd
